Rename builder to match class name
