+++ graphqlapi/src/main/java/io/stargate/graphql/schema/graphqlfirst/processor/DeleteModelBuilder.java
@@ -18,7 +18,7 @@ package io.stargate.graphql.schema.graphqlfirst.processor;
 import graphql.language.Directive;
 import graphql.language.FieldDefinition;
 import graphql.language.InputValueDefinition;
-import io.stargate.graphql.schema.graphqlfirst.processor.ArgumentsDirectiveModelsBuilder.OperationType;
+import io.stargate.graphql.schema.graphqlfirst.processor.ArgumentDirectiveModelsBuilder.OperationType;
 import io.stargate.graphql.schema.graphqlfirst.processor.OperationModel.ReturnType;
 import io.stargate.graphql.schema.graphqlfirst.processor.OperationModel.SimpleReturnType;
 import io.stargate.graphql.schema.graphqlfirst.processor.ResponsePayloadModel.TechnicalField;
@@ -100,7 +100,7 @@ class DeleteModelBuilder extends MutationModelBuilder {
     } else {
       entity = entityFromDirective(cqlDeleteDirective, "delete", CqlDirectives.DELETE);
       ArgumentDirectiveModels conditions =
-          new ArgumentsDirectiveModelsBuilder(
+          new ArgumentDirectiveModelsBuilder(
                   operation, OperationType.DELETE, entity, entities, context)
               .build();
       whereConditions = conditions.getWhereConditions();
