+++ graphqlapi/src/main/java/io/stargate/graphql/schema/graphqlfirst/processor/QueryModelBuilder.java
@@ -21,7 +21,7 @@ import graphql.language.FieldDefinition;
 import graphql.language.InputValueDefinition;
 import graphql.language.Type;
 import graphql.language.TypeName;
-import io.stargate.graphql.schema.graphqlfirst.processor.ArgumentsDirectiveModelsBuilder.OperationType;
+import io.stargate.graphql.schema.graphqlfirst.processor.ArgumentDirectiveModelsBuilder.OperationType;
 import io.stargate.graphql.schema.graphqlfirst.processor.OperationModel.ReturnType;
 import io.stargate.graphql.schema.graphqlfirst.util.TypeHelper;
 import java.util.List;
@@ -74,7 +74,7 @@ class QueryModelBuilder extends OperationModelBuilderBase<QueryModel> {
 
     Optional<String> pagingStateArgumentName = findPagingState();
     ArgumentDirectiveModels conditions =
-        new ArgumentsDirectiveModelsBuilder(
+        new ArgumentDirectiveModelsBuilder(
                 operation, OperationType.SELECT, entity, entities, context)
             .build();
     List<ConditionModel> whereConditions = conditions.getWhereConditions();
